Renamed tasks app to todo app
