--- hypha/apply/tasks/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class StreamFormsConfig(AppConfig):
-    name = "hypha.apply.tasks"
