+++ hypha/apply/todo/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class TodoConfig(AppConfig):
+    name = "hypha.apply.todo"
